+++ package.json
@@ -1,11 +1,12 @@
 {
-  "name": "iox",
+  "name": "ioperator",
   "version": "1.0.0",
-  "main": "lib/iox.js",
-  "repository": "git@github.com:pokle/iox.git",
+  "main": "lib/ioperator.js",
+  "repository": "git@github.com:pokle/ioperator.git",
   "author": "Tushar Pokle <tushar.pokle@gmail.com>",
   "license": "MIT",
   "scripts": {
+    "clean": "rm -rf coverage lib",
     "build": "rm -rf lib && babel src -d lib",
     "test": "jest --coverage"
   },
