+++ src/ioperator.js
